bot/requirements.txt
crewai==0.22.5
flask==3.0.3
flask-cors==5.0.0
langchain-google-genai==1.0.1
python-dotenv==1.0.0
gunicorn==23.0.0